requirements.txt
Django==1.8.4
django-crispy-forms==1.5.2
django-shortuuidfield==0.1.3
mysqlclient==1.3.6
shortuuid==0.4.2
six==1.9.0
